Closes Issue #74 - Restyle Resource pages
Fixed up the padding in the resource carousel and enlarged the height to fit more text.
Added special styling for links in each carousel
Moved the go back button up so it sits always under the resource definition carousel.
Links can be clickable as long as an anchor tag is included in the db as html.